pyproject.toml
[tool.poetry]
name = "django-request-token"
version = "0.12"
description = "JWT-backed Django app for managing querystring tokens."
license = "MIT"
authors = ["YunoJuno <code@yunojuno.com>"]
maintainers = ["YunoJuno <code@yunojuno.com>"]
readme = "README.rst"
homepage = "https://github.com/yunojuno/django-request-token"
repository = "https://github.com/yunojuno/django-request-token"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
]
packages = [{ include = "request_token" }]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2 || ^3.0"
pyjwt = "*"
sqlparse = "*"
psycopg2-binary = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
coverage = "*"
pylint = "*"
pylint-django = "*"
flake8 = "*"
flake8-bandit = "*"
flake8-docstrings = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
black = "=19.10b0"
tox = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
